Changed utterences color scheme and jotted down some notes about github comment schemes.
Advantages of Giscus: It is more appropriate to use the discussions tab than it is to use the Issues tab. Code is actually legible.
Disadvantages of Giscus: Right side of frame is clipped off. Bottom of comment box is clipped off in mobile browser.

Advantages of Utterances: Illegible Code. Broader userbase and support.
Disadvantages of Utterances: Bottom of comment box is clipped off in mobile browser.